Move test functions to public file
